Performs a differential item functioning analysis of polytomous items using ANCOVA

Differential item functioning analysis of polytomous items via ANCOVA

Usage

ANCOVA.dif(x, Group, Trait)

Arguments

x

A data frame containing the items. All the items should be scored in the same direction.

Group

A factor indicating group membership

Trait

A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.

Examples

# Differential item functioning across gender
## Calculate the Trait score
Trait <- rowSums(work_stress[1:9])

#### Perform the DIF analysis
out <- ANCOVA.dif(work_stress[1:9], work_stress$gender, Trait)
out
#> $Results
#>   Item Trait_p Group_p  TxG_p Trait_eta.sq Group_eta.sq TxG_eta.sq
#> 1 GWS1       0 0.26265 0.4402       0.6248     0.000343   0.000163
#> 2 GWS2       0 0.32465 0.2101       0.6199     0.000268   0.000434
#> 3 GWS3       0 0.01207 0.6084       0.6262     0.001711   0.000071
#> 4 GWS4       0 0.18434 0.9555       0.6141     0.000495   0.000001
#> 5 GWS5       0 0.20207 0.7887       0.5972     0.000477   0.000021
#> 6 GWS6       0 0.02527 0.8207       0.5938     0.001479   0.000015
#> 7 GWS7       0 0.12598 0.5350       0.6012     0.000680   0.000112
#> 8 GWS8       0 0.59393 0.8373       0.5990     0.000083   0.000012
#> 9 GWS9       0 0.64020 0.1100       0.5763     0.000067   0.000788
#> 
#> $Most.uniform.dif
#>   Item Trait_eta.sq Group_eta.sq TxG_eta.sq
#> 3 GWS3       0.6262     0.001711 0.00007116
#> 
#> $Most.non.uniform.dif
#>   Item Trait_eta.sq Group_eta.sq TxG_eta.sq
#> 9 GWS9       0.5763   0.00006732  0.0007877
#>