Rasch analysis of dichotomous or polytomous items

Basic Rasch analysis of dichotomous or polytomous items. This function is a wrapper of the tam.mml() function of the TAM package and of the principal() function of the psych package. The purpose of the function is to simplify the Rasch analysis process and to produce the output and plots that applied and beginner users most typically want.

Usage

Simple.Rasch(raschdata, irtmodel)

Arguments

raschdata

A data frame containing items where the lowest response category is scored as 0

irtmodel

The Rasch model, which can be "PCM", "PCM2" or "RSM".

Value

The output is a list containing the results of the analysis.

See also

Examples

# Lowest response category must be scored as 0
work_stress0 <- work_stress[1:9] - 1
# Fit the rating scale model
out <- Simple.Rasch(work_stress0, "RSM")
#> Error in library(psych): there is no package called ‘psych’
ls(out)
#> Warning: ‘out’ converted to character string
#> Error in as.environment(pos): no item called "out" on the search list
out$item.fit
#> Error: object 'out' not found

# Fit the partial credit model
out <- Simple.Rasch(work_stress0, "PCM2")
#> Error in library(psych): there is no package called ‘psych’
ls(out)
#> Warning: ‘out’ converted to character string
#> Error in as.environment(pos): no item called "out" on the search list
out$item.fit
#> Error: object 'out' not found