Fits all possible combinations of k items from a pool of n items to a single-factor model.

The function can be useful in selecting items from a larger pool of items. The function allows for comparisons of all possible combinations of k items from the pool with respect to fit with a single factor model and reliability. The function works best if the number of possible unique combinations of items is less than 5000. It becomes extremely slow if there are more than 5000 possible combinations.

Usage

best_items(x, k)

Arguments

x

A data frame containing the items.

k

The number of items to be selected.

Examples

# Say we want to select five items from pool of nine stress items
# The choose() function shows how many possible unique combinations there are
choose(ncol(work_stress[1:9]), 5)
#> [1] 126
out <- best_items(work_stress[1:9], 5)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
out
#>     Omega Alpha  SRMR   CFI   TLI RMSEA   GFI                        Items
#> 78  0.864 0.864 0.004 1.000 1.003 0.000 1.000 GWS2, GWS3, GWS4, GWS7, GWS8
#> 94  0.864 0.864 0.004 1.000 1.003 0.000 1.000 GWS2, GWS4, GWS5, GWS7, GWS8
#> 4   0.865 0.865 0.005 1.000 1.002 0.000 1.000 GWS1, GWS2, GWS3, GWS4, GWS8
#> 72  0.864 0.864 0.005 1.000 1.002 0.000 1.000 GWS2, GWS3, GWS4, GWS5, GWS7
#> 73  0.867 0.867 0.006 1.000 1.001 0.000 1.000 GWS2, GWS3, GWS4, GWS5, GWS8
#> 18  0.865 0.865 0.006 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS4, GWS5, GWS8
#> 106 0.864 0.864 0.006 1.000 1.001 0.000 1.000 GWS3, GWS4, GWS5, GWS6, GWS7
#> 114 0.863 0.863 0.006 1.000 1.001 0.000 1.000 GWS3, GWS4, GWS6, GWS8, GWS9
#> 45  0.862 0.862 0.006 1.000 1.002 0.000 1.000 GWS1, GWS3, GWS4, GWS8, GWS9
#> 117 0.861 0.860 0.006 1.000 1.002 0.000 1.000 GWS3, GWS5, GWS6, GWS7, GWS9
#> 91  0.860 0.860 0.006 1.000 1.002 0.000 1.000 GWS2, GWS4, GWS5, GWS6, GWS7
#> 8   0.867 0.867 0.007 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS3, GWS5, GWS8
#> 1   0.866 0.866 0.007 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS3, GWS4, GWS5
#> 38  0.866 0.866 0.007 1.000 1.001 0.000 1.000 GWS1, GWS3, GWS4, GWS5, GWS8
#> 84  0.866 0.866 0.007 1.000 1.001 0.000 1.000 GWS2, GWS3, GWS5, GWS7, GWS8
#> 41  0.865 0.865 0.007 1.000 1.001 0.000 1.000 GWS1, GWS3, GWS4, GWS6, GWS8
#> 42  0.865 0.865 0.007 1.000 1.001 0.000 1.000 GWS1, GWS3, GWS4, GWS6, GWS9
#> 109 0.865 0.865 0.007 1.000 1.001 0.000 1.000 GWS3, GWS4, GWS5, GWS7, GWS8
#> 5   0.864 0.864 0.007 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS3, GWS4, GWS9
#> 15  0.863 0.863 0.007 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS3, GWS8, GWS9
#> 54  0.863 0.863 0.007 1.000 1.001 0.000 1.000 GWS1, GWS3, GWS6, GWS8, GWS9
#> 92  0.863 0.863 0.007 1.000 1.001 0.000 1.000 GWS2, GWS4, GWS5, GWS6, GWS8
#> 112 0.863 0.863 0.007 1.000 1.001 0.000 1.000 GWS3, GWS4, GWS6, GWS7, GWS8
#> 121 0.862 0.862 0.007 1.000 1.001 0.000 1.000 GWS4, GWS5, GWS6, GWS7, GWS8
#> 31  0.861 0.861 0.007 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS5, GWS8, GWS9
#> 58  0.861 0.861 0.007 1.000 1.001 0.000 1.000 GWS1, GWS4, GWS5, GWS6, GWS9
#> 61  0.861 0.861 0.007 1.000 1.001 0.000 1.000 GWS1, GWS4, GWS5, GWS8, GWS9
#> 64  0.861 0.861 0.007 1.000 1.001 0.000 1.000 GWS1, GWS4, GWS6, GWS8, GWS9
#> 101 0.861 0.861 0.007 1.000 1.001 0.000 1.000 GWS2, GWS5, GWS6, GWS7, GWS8
#> 124 0.858 0.858 0.007 1.000 1.001 0.000 1.000 GWS4, GWS5, GWS7, GWS8, GWS9
#> 122 0.857 0.857 0.007 1.000 1.001 0.000 1.000 GWS4, GWS5, GWS6, GWS7, GWS9
#> 20  0.864 0.864 0.008 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS4, GWS6, GWS7
#> 71  0.864 0.864 0.008 1.000 1.001 0.000 1.000 GWS2, GWS3, GWS4, GWS5, GWS6
#> 80  0.864 0.863 0.008 1.000 1.001 0.000 1.000 GWS2, GWS3, GWS4, GWS8, GWS9
#> 48  0.863 0.863 0.008 1.000 1.000 0.003 1.000 GWS1, GWS3, GWS5, GWS6, GWS9
#> 77  0.863 0.863 0.008 1.000 1.000 0.000 1.000 GWS2, GWS3, GWS4, GWS6, GWS9
#> 81  0.863 0.863 0.008 1.000 1.001 0.000 1.000 GWS2, GWS3, GWS5, GWS6, GWS7
#> 25  0.862 0.862 0.008 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS4, GWS8, GWS9
#> 75  0.862 0.862 0.008 1.000 1.001 0.000 1.000 GWS2, GWS3, GWS4, GWS6, GWS7
#> 19  0.861 0.861 0.008 1.000 1.001 0.000 1.000 GWS1, GWS2, GWS4, GWS5, GWS9
#> 97  0.861 0.861 0.008 1.000 1.001 0.000 1.000 GWS2, GWS4, GWS6, GWS7, GWS8
#> 113 0.861 0.861 0.008 1.000 1.001 0.000 1.000 GWS3, GWS4, GWS6, GWS7, GWS9
#> 104 0.859 0.859 0.008 1.000 1.001 0.000 1.000 GWS2, GWS5, GWS7, GWS8, GWS9
#> 36  0.867 0.867 0.009 1.000 0.999 0.013 1.000 GWS1, GWS3, GWS4, GWS5, GWS6
#> 43  0.866 0.866 0.009 1.000 1.000 0.010 1.000 GWS1, GWS3, GWS4, GWS7, GWS8
#> 107 0.866 0.866 0.009 1.000 1.000 0.007 1.000 GWS3, GWS4, GWS5, GWS6, GWS8
#> 2   0.865 0.865 0.009 0.999 0.999 0.015 1.000 GWS1, GWS2, GWS3, GWS4, GWS6
#> 3   0.865 0.865 0.009 1.000 1.000 0.008 1.000 GWS1, GWS2, GWS3, GWS4, GWS7
#> 62  0.865 0.865 0.009 1.000 0.999 0.011 1.000 GWS1, GWS4, GWS6, GWS7, GWS8
#> 57  0.864 0.864 0.009 1.000 0.999 0.012 1.000 GWS1, GWS4, GWS5, GWS6, GWS8
#> 108 0.864 0.864 0.009 1.000 1.000 0.009 1.000 GWS3, GWS4, GWS5, GWS6, GWS9
#> 116 0.864 0.864 0.009 1.000 0.999 0.014 1.000 GWS3, GWS5, GWS6, GWS7, GWS8
#> 9   0.863 0.863 0.009 1.000 1.000 0.007 1.000 GWS1, GWS2, GWS3, GWS5, GWS9
#> 16  0.863 0.863 0.009 1.000 1.000 0.006 1.000 GWS1, GWS2, GWS4, GWS5, GWS6
#> 39  0.863 0.863 0.009 1.000 0.999 0.012 1.000 GWS1, GWS3, GWS4, GWS5, GWS9
#> 51  0.862 0.862 0.009 1.000 0.999 0.012 1.000 GWS1, GWS3, GWS5, GWS8, GWS9
#> 79  0.861 0.861 0.009 1.000 1.000 0.006 1.000 GWS2, GWS3, GWS4, GWS7, GWS9
#> 85  0.861 0.861 0.009 1.000 0.999 0.012 1.000 GWS2, GWS3, GWS5, GWS7, GWS9
#> 110 0.860 0.860 0.009 1.000 0.999 0.012 1.000 GWS3, GWS4, GWS5, GWS7, GWS9
#> 123 0.860 0.860 0.009 1.000 1.000 0.008 1.000 GWS4, GWS5, GWS6, GWS8, GWS9
#> 95  0.858 0.858 0.009 1.000 1.000 0.007 1.000 GWS2, GWS4, GWS5, GWS7, GWS9
#> 125 0.858 0.858 0.009 1.000 1.000 0.006 1.000 GWS4, GWS6, GWS7, GWS8, GWS9
#> 40  0.867 0.867 0.010 0.999 0.998 0.019 1.000 GWS1, GWS3, GWS4, GWS6, GWS7
#> 13  0.866 0.866 0.010 0.999 0.998 0.020 1.000 GWS1, GWS2, GWS3, GWS7, GWS8
#> 23  0.866 0.866 0.010 0.999 0.998 0.021 1.000 GWS1, GWS2, GWS4, GWS7, GWS8
#> 21  0.864 0.864 0.010 0.999 0.998 0.019 1.000 GWS1, GWS2, GWS4, GWS6, GWS8
#> 56  0.864 0.864 0.010 0.999 0.999 0.018 1.000 GWS1, GWS4, GWS5, GWS6, GWS7
#> 74  0.864 0.864 0.010 0.999 0.999 0.016 1.000 GWS2, GWS3, GWS4, GWS5, GWS9
#> 111 0.863 0.863 0.010 0.999 0.998 0.019 1.000 GWS3, GWS4, GWS5, GWS8, GWS9
#> 118 0.863 0.863 0.010 0.999 0.998 0.019 1.000 GWS3, GWS5, GWS6, GWS8, GWS9
#> 90  0.862 0.861 0.010 0.999 0.999 0.017 1.000 GWS2, GWS3, GWS7, GWS8, GWS9
#> 96  0.862 0.862 0.010 0.999 0.999 0.016 1.000 GWS2, GWS4, GWS5, GWS8, GWS9
#> 93  0.860 0.860 0.010 0.999 0.999 0.016 1.000 GWS2, GWS4, GWS5, GWS6, GWS9
#> 99  0.860 0.860 0.010 0.999 0.998 0.018 1.000 GWS2, GWS4, GWS6, GWS8, GWS9
#> 100 0.860 0.860 0.010 1.000 0.999 0.013 1.000 GWS2, GWS4, GWS7, GWS8, GWS9
#> 115 0.860 0.860 0.010 1.000 0.999 0.013 1.000 GWS3, GWS4, GWS7, GWS8, GWS9
#> 119 0.860 0.860 0.010 0.999 0.998 0.018 1.000 GWS3, GWS5, GWS7, GWS8, GWS9
#> 120 0.860 0.860 0.010 1.000 0.999 0.012 1.000 GWS3, GWS6, GWS7, GWS8, GWS9
#> 126 0.857 0.857 0.010 0.999 0.999 0.016 1.000 GWS5, GWS6, GWS7, GWS8, GWS9
#> 47  0.866 0.866 0.011 0.999 0.997 0.024 1.000 GWS1, GWS3, GWS5, GWS6, GWS8
#> 52  0.866 0.866 0.011 0.999 0.998 0.022 1.000 GWS1, GWS3, GWS6, GWS7, GWS8
#> 6   0.865 0.865 0.011 0.999 0.997 0.024 1.000 GWS1, GWS2, GWS3, GWS5, GWS6
#> 32  0.865 0.864 0.011 0.999 0.997 0.025 1.000 GWS1, GWS2, GWS6, GWS7, GWS8
#> 82  0.865 0.865 0.011 0.999 0.998 0.020 1.000 GWS2, GWS3, GWS5, GWS6, GWS8
#> 17  0.864 0.864 0.011 0.999 0.998 0.021 1.000 GWS1, GWS2, GWS4, GWS5, GWS7
#> 76  0.864 0.864 0.011 0.999 0.998 0.022 1.000 GWS2, GWS3, GWS4, GWS6, GWS8
#> 86  0.864 0.864 0.011 0.999 0.997 0.024 1.000 GWS2, GWS3, GWS5, GWS8, GWS9
#> 26  0.863 0.863 0.011 0.999 0.998 0.020 1.000 GWS1, GWS2, GWS5, GWS6, GWS7
#> 83  0.863 0.863 0.011 0.999 0.998 0.021 1.000 GWS2, GWS3, GWS5, GWS6, GWS9
#> 87  0.863 0.863 0.011 0.999 0.998 0.020 1.000 GWS2, GWS3, GWS6, GWS7, GWS8
#> 22  0.862 0.862 0.011 0.999 0.998 0.022 1.000 GWS1, GWS2, GWS4, GWS6, GWS9
#> 68  0.860 0.860 0.011 0.999 0.998 0.019 1.000 GWS1, GWS5, GWS6, GWS8, GWS9
#> 67  0.859 0.859 0.011 0.999 0.997 0.023 1.000 GWS1, GWS5, GWS6, GWS7, GWS9
#> 98  0.858 0.858 0.011 0.999 0.997 0.023 1.000 GWS2, GWS4, GWS6, GWS7, GWS9
#> 10  0.866 0.866 0.012 0.998 0.996 0.028 1.000 GWS1, GWS2, GWS3, GWS6, GWS7
#> 12  0.864 0.864 0.012 0.998 0.996 0.028 1.000 GWS1, GWS2, GWS3, GWS6, GWS9
#> 27  0.863 0.863 0.012 0.998 0.997 0.026 1.000 GWS1, GWS2, GWS5, GWS6, GWS8
#> 89  0.863 0.863 0.012 0.998 0.997 0.025 1.000 GWS2, GWS3, GWS6, GWS8, GWS9
#> 88  0.861 0.861 0.012 0.999 0.997 0.024 1.000 GWS2, GWS3, GWS6, GWS7, GWS9
#> 28  0.860 0.860 0.012 0.998 0.996 0.028 1.000 GWS1, GWS2, GWS5, GWS6, GWS9
#> 70  0.860 0.860 0.012 0.998 0.996 0.027 1.000 GWS1, GWS6, GWS7, GWS8, GWS9
#> 103 0.860 0.860 0.012 0.998 0.996 0.028 0.999 GWS2, GWS5, GWS6, GWS8, GWS9
#> 102 0.856 0.856 0.012 0.998 0.996 0.026 1.000 GWS2, GWS5, GWS6, GWS7, GWS9
#> 7   0.866 0.866 0.013 0.998 0.995 0.032 0.999 GWS1, GWS2, GWS3, GWS5, GWS7
#> 29  0.866 0.866 0.013 0.998 0.996 0.031 0.999 GWS1, GWS2, GWS5, GWS7, GWS8
#> 37  0.866 0.866 0.013 0.997 0.995 0.033 0.999 GWS1, GWS3, GWS4, GWS5, GWS7
#> 11  0.865 0.865 0.013 0.998 0.996 0.030 1.000 GWS1, GWS2, GWS3, GWS6, GWS8
#> 59  0.865 0.865 0.013 0.998 0.996 0.030 0.999 GWS1, GWS4, GWS5, GWS7, GWS8
#> 66  0.864 0.864 0.013 0.998 0.995 0.031 0.999 GWS1, GWS5, GWS6, GWS7, GWS8
#> 63  0.862 0.862 0.013 0.998 0.996 0.029 0.999 GWS1, GWS4, GWS6, GWS7, GWS9
#> 34  0.861 0.861 0.013 0.998 0.995 0.031 0.999 GWS1, GWS2, GWS6, GWS8, GWS9
#> 65  0.861 0.861 0.013 0.998 0.996 0.029 0.999 GWS1, GWS4, GWS7, GWS8, GWS9
#> 60  0.859 0.859 0.013 0.998 0.996 0.030 0.999 GWS1, GWS4, GWS5, GWS7, GWS9
#> 69  0.859 0.859 0.013 0.998 0.995 0.031 0.999 GWS1, GWS5, GWS7, GWS8, GWS9
#> 105 0.858 0.858 0.013 0.998 0.995 0.031 0.999 GWS2, GWS6, GWS7, GWS8, GWS9
#> 46  0.867 0.867 0.014 0.997 0.994 0.035 0.999 GWS1, GWS3, GWS5, GWS6, GWS7
#> 49  0.866 0.866 0.014 0.997 0.994 0.036 0.999 GWS1, GWS3, GWS5, GWS7, GWS8
#> 35  0.862 0.862 0.014 0.997 0.994 0.035 0.999 GWS1, GWS2, GWS7, GWS8, GWS9
#> 55  0.861 0.861 0.014 0.997 0.994 0.035 0.999 GWS1, GWS3, GWS7, GWS8, GWS9
#> 30  0.860 0.859 0.014 0.997 0.995 0.033 0.999 GWS1, GWS2, GWS5, GWS7, GWS9
#> 53  0.864 0.864 0.015 0.997 0.993 0.037 0.999 GWS1, GWS3, GWS6, GWS7, GWS9
#> 24  0.862 0.861 0.015 0.997 0.993 0.037 0.999 GWS1, GWS2, GWS4, GWS7, GWS9
#> 44  0.862 0.862 0.015 0.997 0.993 0.038 0.999 GWS1, GWS3, GWS4, GWS7, GWS9
#> 50  0.861 0.861 0.015 0.996 0.992 0.041 0.999 GWS1, GWS3, GWS5, GWS7, GWS9
#> 14  0.863 0.863 0.016 0.996 0.992 0.040 0.999 GWS1, GWS2, GWS3, GWS7, GWS9
#> 33  0.861 0.861 0.016 0.996 0.991 0.042 0.999 GWS1, GWS2, GWS6, GWS7, GWS9