Inter-item correlations

Statistics related to the inter-item correlations of the items

cor.item(x)

Arguments

x

Data frame containing items that are all scored in the same direction

Value

A list containing (a) the inter-item correlation matrix, (b) for each item its average correlation with the remaining items, (c) the squared multiple correlation of each item with the remaining items, and (d) the average inter-item correlation across all the items, and (e) the median inter-item correlation across all the items.

Examples

library(lordif)
data(Anxiety)
cor.item(Anxiety[4:32])
#> $Correlation.matrix
#>        R1    R2    R3    R4    R5    R6    R7    R8    R9   R10   R11   R12
#> R1  1.000 0.781 0.723 0.642 0.567 0.618 0.624 0.434 0.484 0.691 0.510 0.503
#> R2  0.781 1.000 0.713 0.625 0.599 0.611 0.591 0.454 0.457 0.707 0.526 0.491
#> R3  0.723 0.713 1.000 0.650 0.638 0.674 0.570 0.504 0.482 0.734 0.495 0.520
#> R4  0.642 0.625 0.650 1.000 0.713 0.598 0.657 0.474 0.533 0.695 0.519 0.562
#> R5  0.567 0.599 0.638 0.713 1.000 0.650 0.542 0.458 0.496 0.679 0.493 0.479
#> R6  0.618 0.611 0.674 0.598 0.650 1.000 0.534 0.409 0.448 0.641 0.446 0.536
#> R7  0.624 0.591 0.570 0.657 0.542 0.534 1.000 0.407 0.464 0.591 0.469 0.524
#> R8  0.434 0.454 0.504 0.474 0.458 0.409 0.407 1.000 0.371 0.530 0.391 0.390
#> R9  0.484 0.457 0.482 0.533 0.496 0.448 0.464 0.371 1.000 0.469 0.478 0.505
#> R10 0.691 0.707 0.734 0.695 0.679 0.641 0.591 0.530 0.469 1.000 0.510 0.486
#> R11 0.510 0.526 0.495 0.519 0.493 0.446 0.469 0.391 0.478 0.510 1.000 0.532
#> R12 0.503 0.491 0.520 0.562 0.479 0.536 0.524 0.390 0.505 0.486 0.532 1.000
#> R13 0.504 0.420 0.494 0.528 0.491 0.506 0.495 0.351 0.530 0.532 0.411 0.495
#> R14 0.489 0.486 0.509 0.583 0.521 0.486 0.483 0.409 0.455 0.522 0.448 0.554
#> R15 0.713 0.683 0.659 0.634 0.556 0.608 0.572 0.448 0.462 0.688 0.515 0.492
#> R16 0.646 0.618 0.579 0.682 0.570 0.584 0.694 0.415 0.498 0.601 0.475 0.572
#> R17 0.705 0.735 0.683 0.562 0.573 0.603 0.499 0.427 0.432 0.705 0.502 0.435
#> R18 0.532 0.491 0.513 0.539 0.470 0.544 0.535 0.310 0.455 0.495 0.416 0.530
#> R19 0.650 0.622 0.691 0.696 0.702 0.595 0.575 0.494 0.555 0.709 0.499 0.551
#> R20 0.688 0.666 0.656 0.640 0.640 0.614 0.613 0.456 0.512 0.691 0.519 0.521
#> R21 0.394 0.382 0.454 0.420 0.428 0.402 0.354 0.386 0.352 0.493 0.329 0.388
#> R22 0.661 0.639 0.669 0.756 0.619 0.576 0.638 0.507 0.528 0.676 0.522 0.575
#> R23 0.536 0.523 0.540 0.604 0.549 0.577 0.600 0.349 0.497 0.546 0.463 0.638
#> R24 0.641 0.583 0.590 0.635 0.571 0.575 0.648 0.425 0.539 0.620 0.511 0.536
#> R25 0.394 0.372 0.382 0.471 0.425 0.403 0.441 0.362 0.370 0.394 0.376 0.444
#> R26 0.550 0.517 0.535 0.620 0.543 0.527 0.604 0.412 0.476 0.564 0.500 0.520
#> R27 0.679 0.629 0.644 0.724 0.617 0.593 0.687 0.488 0.533 0.656 0.509 0.580
#> R28 0.604 0.586 0.566 0.676 0.550 0.566 0.667 0.432 0.483 0.585 0.479 0.576
#> R29 0.633 0.604 0.647 0.689 0.637 0.585 0.645 0.521 0.544 0.712 0.528 0.572
#>       R13   R14   R15   R16   R17   R18   R19   R20   R21   R22   R23   R24
#> R1  0.504 0.489 0.713 0.646 0.705 0.532 0.650 0.688 0.394 0.661 0.536 0.641
#> R2  0.420 0.486 0.683 0.618 0.735 0.491 0.622 0.666 0.382 0.639 0.523 0.583
#> R3  0.494 0.509 0.659 0.579 0.683 0.513 0.691 0.656 0.454 0.669 0.540 0.590
#> R4  0.528 0.583 0.634 0.682 0.562 0.539 0.696 0.640 0.420 0.756 0.604 0.635
#> R5  0.491 0.521 0.556 0.570 0.573 0.470 0.702 0.640 0.428 0.619 0.549 0.571
#> R6  0.506 0.486 0.608 0.584 0.603 0.544 0.595 0.614 0.402 0.576 0.577 0.575
#> R7  0.495 0.483 0.572 0.694 0.499 0.535 0.575 0.613 0.354 0.638 0.600 0.648
#> R8  0.351 0.409 0.448 0.415 0.427 0.310 0.494 0.456 0.386 0.507 0.349 0.425
#> R9  0.530 0.455 0.462 0.498 0.432 0.455 0.555 0.512 0.352 0.528 0.497 0.539
#> R10 0.532 0.522 0.688 0.601 0.705 0.495 0.709 0.691 0.493 0.676 0.546 0.620
#> R11 0.411 0.448 0.515 0.475 0.502 0.416 0.499 0.519 0.329 0.522 0.463 0.511
#> R12 0.495 0.554 0.492 0.572 0.435 0.530 0.551 0.521 0.388 0.575 0.638 0.536
#> R13 1.000 0.456 0.503 0.524 0.449 0.509 0.537 0.543 0.385 0.569 0.513 0.567
#> R14 0.456 1.000 0.451 0.544 0.452 0.473 0.514 0.496 0.411 0.597 0.536 0.511
#> R15 0.503 0.451 1.000 0.584 0.679 0.521 0.639 0.657 0.395 0.608 0.501 0.610
#> R16 0.524 0.544 0.584 1.000 0.505 0.556 0.603 0.670 0.361 0.705 0.609 0.717
#> R17 0.449 0.452 0.679 0.505 1.000 0.468 0.616 0.622 0.378 0.557 0.452 0.533
#> R18 0.509 0.473 0.521 0.556 0.468 1.000 0.462 0.534 0.295 0.571 0.557 0.560
#> R19 0.537 0.514 0.639 0.603 0.616 0.462 1.000 0.689 0.443 0.660 0.581 0.626
#> R20 0.543 0.496 0.657 0.670 0.622 0.534 0.689 1.000 0.392 0.641 0.566 0.690
#> R21 0.385 0.411 0.395 0.361 0.378 0.295 0.443 0.392 1.000 0.425 0.389 0.389
#> R22 0.569 0.597 0.608 0.705 0.557 0.571 0.660 0.641 0.425 1.000 0.619 0.656
#> R23 0.513 0.536 0.501 0.609 0.452 0.557 0.581 0.566 0.389 0.619 1.000 0.567
#> R24 0.567 0.511 0.610 0.717 0.533 0.560 0.626 0.690 0.389 0.656 0.567 1.000
#> R25 0.378 0.420 0.351 0.485 0.351 0.323 0.389 0.416 0.341 0.452 0.429 0.454
#> R26 0.483 0.554 0.495 0.628 0.460 0.490 0.570 0.576 0.386 0.630 0.539 0.611
#> R27 0.555 0.593 0.603 0.736 0.562 0.610 0.643 0.631 0.391 0.744 0.655 0.690
#> R28 0.500 0.554 0.557 0.679 0.481 0.558 0.577 0.592 0.402 0.680 0.593 0.634
#> R29 0.564 0.565 0.626 0.622 0.584 0.520 0.661 0.660 0.423 0.678 0.579 0.649
#>       R25   R26   R27   R28   R29
#> R1  0.394 0.550 0.679 0.604 0.633
#> R2  0.372 0.517 0.629 0.586 0.604
#> R3  0.382 0.535 0.644 0.566 0.647
#> R4  0.471 0.620 0.724 0.676 0.689
#> R5  0.425 0.543 0.617 0.550 0.637
#> R6  0.403 0.527 0.593 0.566 0.585
#> R7  0.441 0.604 0.687 0.667 0.645
#> R8  0.362 0.412 0.488 0.432 0.521
#> R9  0.370 0.476 0.533 0.483 0.544
#> R10 0.394 0.564 0.656 0.585 0.712
#> R11 0.376 0.500 0.509 0.479 0.528
#> R12 0.444 0.520 0.580 0.576 0.572
#> R13 0.378 0.483 0.555 0.500 0.564
#> R14 0.420 0.554 0.593 0.554 0.565
#> R15 0.351 0.495 0.603 0.557 0.626
#> R16 0.485 0.628 0.736 0.679 0.622
#> R17 0.351 0.460 0.562 0.481 0.584
#> R18 0.323 0.490 0.610 0.558 0.520
#> R19 0.389 0.570 0.643 0.577 0.661
#> R20 0.416 0.576 0.631 0.592 0.660
#> R21 0.341 0.386 0.391 0.402 0.423
#> R22 0.452 0.630 0.744 0.680 0.678
#> R23 0.429 0.539 0.655 0.593 0.579
#> R24 0.454 0.611 0.690 0.634 0.649
#> R25 1.000 0.597 0.436 0.443 0.438
#> R26 0.597 1.000 0.628 0.651 0.623
#> R27 0.436 0.628 1.000 0.704 0.666
#> R28 0.443 0.651 0.704 1.000 0.646
#> R29 0.438 0.623 0.666 0.646 1.000
#> 
#> $Average.r
#> [1] 0.545
#> 
#> $Median.r
#> [1] 0.543
#> 
#> $Item.average.r
#>    R1    R2    R3    R4    R5    R6    R7    R8    R9   R10   R11   R12   R13 
#> 0.593 0.575 0.590 0.612 0.563 0.554 0.561 0.429 0.479 0.604 0.478 0.518 0.493 
#>   R14   R15   R16   R17   R18   R19   R20   R21   R22   R23   R24   R25   R26 
#> 0.503 0.565 0.588 0.536 0.494 0.591 0.592 0.392 0.613 0.539 0.584 0.412 0.546 
#>   R27   R28   R29 
#> 0.614 0.572 0.601 
#> 
#> $Item.average.r.sorted
#>   R21   R25    R8   R11    R9   R13   R18   R14   R12   R17   R23   R26    R6 
#> 0.392 0.412 0.429 0.478 0.479 0.493 0.494 0.503 0.518 0.536 0.539 0.546 0.554 
#>    R7    R5   R15   R28    R2   R24   R16    R3   R19   R20    R1   R29   R10 
#> 0.561 0.563 0.565 0.572 0.575 0.584 0.588 0.590 0.591 0.592 0.593 0.601 0.604 
#>    R4   R22   R27 
#> 0.612 0.613 0.614 
#> 
#> $Squared.multiple.correlation
#>    R1    R2    R3    R4    R5    R6    R7    R8    R9   R10   R11   R12   R13 
#> 0.736 0.730 0.702 0.730 0.662 0.614 0.628 0.391 0.459 0.739 0.449 0.571 0.492 
#>   R14   R15   R16   R17   R18   R19   R20   R21   R22   R23   R24   R25   R26 
#> 0.496 0.650 0.703 0.667 0.507 0.692 0.676 0.328 0.719 0.594 0.662 0.424 0.621 
#>   R27   R28   R29 
#> 0.733 0.649 0.677 
#> 
#> $Squared.multiple.correlation.sorted
#>   R21    R8   R25   R11    R9   R13   R14   R18   R12   R23    R6   R26    R7 
#> 0.328 0.391 0.424 0.449 0.459 0.492 0.496 0.507 0.571 0.594 0.614 0.621 0.628 
#>   R28   R15   R24    R5   R17   R20   R29   R19    R3   R16   R22    R2    R4 
#> 0.649 0.650 0.662 0.662 0.667 0.676 0.677 0.692 0.702 0.703 0.719 0.730 0.730 
#>   R27    R1   R10 
#> 0.733 0.736 0.739 
#> 
#> $Item.total.correlation
#>    R1    R2    R3    R4    R5    R6    R7    R8    R9   R10   R11   R12   R13 
#> 0.802 0.777 0.797 0.837 0.769 0.758 0.775 0.593 0.666 0.815 0.661 0.720 0.687 
#>   R14   R15   R16   R17   R18   R19   R20   R21   R22   R23   R24   R25   R26 
#> 0.699 0.766 0.811 0.721 0.690 0.801 0.806 0.549 0.838 0.749 0.803 0.591 0.760 
#>   R27   R28   R29 
#> 0.841 0.791 0.819 
#> 
#> $Item.total.correlation.sorted
#>   R21   R25    R8   R11    R9   R13   R18   R14   R12   R17   R23    R6   R26 
#> 0.549 0.591 0.593 0.661 0.666 0.687 0.690 0.699 0.720 0.721 0.749 0.758 0.760 
#>   R15    R5    R7    R2   R28    R3   R19    R1   R24   R20   R16   R10   R29 
#> 0.766 0.769 0.775 0.777 0.791 0.797 0.801 0.802 0.803 0.806 0.811 0.815 0.819 
#>    R4   R22   R27 
#> 0.837 0.838 0.841 
#>