Detection of statistically significant differential item functioning via ANCOVA of standardized Rasch residuals

Detection of statistically significant differential item functioning via ANCOVA of standardized Rasch residuals

Usage

dif.Rasch(std.residuals, Theta, Group)

Arguments

std.residuals

A matrix or data frame containing the standardized residuals of a Rasch analysis

Theta

A vector containing thetas (person measures)

Group

A vector (factor) containing group membership

Value

DIF results

Examples

library(TAM)
#> Error in library(TAM): there is no package called ‘TAM’
library(lordif)
#> Error in library(lordif): there is no package called ‘lordif’
Anxiety <- Anxiety - 1
#> Error: object 'Anxiety' not found

## Fit the partial credit model
myRasch <- Simple.Rasch(Anxiety[ , 4:32], "PCM2")
#> Error in library(psych): there is no package called ‘psych’

## Perform the DIF analysis using the standardized residuals, Rasch person measure and group as input
dif.Rasch(myRasch$std.residuals, myRasch$thetas[,1], Anxiety$gender)
#> Error: object 'myRasch' not found