Performs a differential item functioning analysis of polytomous items using ANCOVA

Differential item functioning analysis of polytomous items via ANCOVA

Usage

dif.ancova(data, Trait, Group)

Arguments

data

A data frame containing the items. All the items should be scored in the same direction.

Trait

A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.

Group

A factor indicating group membership

Examples

out <- dif.ancova(work_stress[1:9], rowSums(work_stress[1:9]), work_stress$gender)
out
#>      Trait_eta.sq Udif_eta.sq NUdif_eta.sq Trait_p Udif_p NUdif_p
#> GWS1       0.6248      0.0003       0.0002       0 0.2627  0.4402
#> GWS2       0.6199      0.0003       0.0004       0 0.3247  0.2101
#> GWS3       0.6262      0.0017       0.0001       0 0.0121  0.6084
#> GWS4       0.6141      0.0005       0.0000       0 0.1843  0.9555
#> GWS5       0.5972      0.0005       0.0000       0 0.2021  0.7887
#> GWS6       0.5938      0.0015       0.0000       0 0.0253  0.8207
#> GWS7       0.6012      0.0007       0.0001       0 0.1260  0.5350
#> GWS8       0.5990      0.0001       0.0000       0 0.5939  0.8373
#> GWS9       0.5763      0.0001       0.0008       0 0.6402  0.1100