Item selection based on McDonald's item information statistic

An item selection procedure based on item information statistics.

info.select(x, nitems)

Arguments

x

A data frame containing items

nitems

The number of items to retain in the scale

Value

The function sequentially removes the item with the lowest information statistic from a pool of items. The output is a list containing (a) a data frame that shows at each step of the item selection process the number of items that had been removed, the name of the item that is removed, and the reliability of the scale; (b) the names of the items that were retained; (c) coefficient omega of the final scale, and (d) coefficient alpha of the final scale.

Examples

library(lordif)
data(Anxiety)
info.select(Anxiety[4:32], 5)
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Omega_h for 1 factor is not meaningful, just omega_t
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> Warning: Omega_h and Omega_asymptotic are not meaningful with one factor
#> Warning: diag(V) had non-positive or NA entries; the non-finite result may be dubious
#> $Steps
#>    Items.out Item.removed     Omega
#> 1          0         None 0.9724280
#> 2          1          R21 0.9726342
#> 3          2          R25 0.9727744
#> 4          3           R8 0.9728561
#> 5          4          R11 0.9725792
#> 6          5           R9 0.9723392
#> 7          6          R13 0.9720041
#> 8          7          R18 0.9715956
#> 9          8          R14 0.9712478
#> 10         9          R12 0.9708714
#> 11        10          R17 0.9699203
#> 12        11          R23 0.9691579
#> 13        12           R6 0.9680433
#> 14        13          R26 0.9671320
#> 15        14          R15 0.9654658
#> 16        15           R5 0.9638624
#> 17        16           R2 0.9615250
#> 18        17           R7 0.9591962
#> 19        18           R3 0.9561057
#> 20        19          R28 0.9530974
#> 21        20          R19 0.9487334
#> 22        21           R1 0.9435775
#> 23        22          R24 0.9370725
#> 24        23          R20 0.9304054
#> 25        24          R10 0.9215332
#> 
#> $Items.in.scale
#> [1] "R4"  "R16" "R22" "R27" "R29"
#> 
#> $Omega
#> [1] 0.9215332
#> 
#> $Alpha
#>  raw_alpha
#>  0.9199953
#>