Gulliksen's item reliability index extended to polytomous items

Takes the item-total correlation and multiplies it with the item standard deviation

Usage

item.reliability.index(x)

Arguments

x

A data frame that contains the items to be analysed

Value

The item reliability index of each item

Examples

#Calculate item reliability indexes for the items of the Neuroticism Scale of the bfi data set
library(psychTools)
#> 
#> Attaching package: 'psychTools'
#> The following object is masked from 'package:lordif':
#> 
#>     recode
data(bfi)
item.reliability.index(bfi[16:20])
#> $rel.index
#>       N1       N2       N3       N4       N5 
#> 1.256766 1.201351 1.295280 1.122598 1.101666 
#> 
#> $sorted.rel.index
#>       N5       N4       N2       N1       N3 
#> 1.101666 1.122598 1.201351 1.256766 1.295280 
#> 
#> $item.stats
#>       n     raw.r     std.r     r.cor    r.drop     mean       sd rel.index
#> N1 2778 0.8000205 0.8022606 0.7642900 0.6672057 2.929086 1.570917  1.256766
#> N2 2779 0.7872842 0.7914195 0.7491076 0.6526162 3.507737 1.525944  1.201351
#> N3 2789 0.8080845 0.8059858 0.7426061 0.6748167 3.216565 1.602902  1.295280
#> N4 2764 0.7151742 0.7144252 0.5983692 0.5428001 3.185601 1.569685  1.122598
#> N5 2771 0.6806090 0.6743766 0.5318346 0.4864824 2.969686 1.618647  1.101666
#>