Selecting items based on Gulliksen's item reliability index

Iteratively removes item with the lowest item reliability index from a pool of items. At each step this yields a test with the largest possible variance.

item.selection(x, nitems)

Arguments

x

A data frame that contains the items to be analysed

nitems

The number of items to retain. This value should never be less than three.

Value

At each step the number of items that have been removed, the name of the last item that has been removed, omega, the average item score, and the standard deviation of the test.

Examples

#Calculate item reliability indexes for the items of the Neuroticism Scale of the bfi data set
library(psychTools)
data(bfi)
item.selection(bfi[16:20], 3)
#> $Steps
#>   Items.out Items.in Item.removed     Omega     Alpha Average r     Mean
#> 1         0        5         None 0.8005117 0.8139629 0.4676244 3.163920
#> 2         1        4           N5 0.8044315 0.8126022 0.5202316 3.208671
#> 3         2        3           N4 0.8238606 0.8200087 0.6041697 3.214580
#>         SD
#> 1 1.194916
#> 2 1.254795
#> 3 1.343611
#> 
#> $Items.in.scale
#> [1] "N1" "N2" "N3"
#> 
#> $Omega
#> $Omega$est
#> [1] 0.8238606
#> 
#> 
#> $Alpha
#>  raw_alpha
#>  0.8200087
#>