Reverse code items

Reflects the scoring of items

Usage

reverse.code(x, min.score, max.score, items)

Arguments

x

Data frame containing items, some of which may need to be reverse coded

min.score

Minimum possible score on the items

max.score

Maximum possible score on the items

items

A vector of item names or column numbers of items that need to be reverse coded

Value

A data frame

Examples

library(psychTools)
#> Error in library(psychTools): there is no package called ‘psychTools’
data(bfi)
#> Warning: data set ‘bfi’ not found

## Perform principal components analysis to observe the signs of the component loadings
psych::pca(bfi[1:25], 5)
#> Error in loadNamespace(x): there is no package called ‘psych’
items <- c("A1", "C4", "C5", "E1", "E2", "O2", "O5")
bfi.new <- reverse.code(bfi[1:25], min.score = 1, max.score = 6, items)
#> Error: object 'bfi' not found
psych::pca(bfi.new[1:25], 5)
#> Error in loadNamespace(x): there is no package called ‘psych’