Performs a differential item functioning analysis of polytomous items using ANCOVA

Differential item functioning analysis of polytomous items via ANCOVA

Usage

ANCOVA.dif(x, Group, Trait)

Arguments

x

A data frame containing the items. All the items should be scored in the same direction.

Group

A factor indicating group membership

Trait

A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.

Examples

library(lordif)
#> Loading required package: mirt
#> Loading required package: stats4
#> Loading required package: lattice
#> Loading required package: rms
#> Loading required package: Hmisc
#> 
#> Attaching package: 'Hmisc'
#> The following objects are masked from 'package:base':
#> 
#>     format.pval, units
data(Anxiety)

#'### Calculate the Trait score
Trait <- rowMeans(Anxiety[4:32])

#'### Perform the DIF analysis
out <- ANCOVA.dif(Anxiety[4:32], Anxiety$gender, Trait)
out
#> $Results
#>    Item Trait_p  Group_p    TxG_p Trait_eta.sq Group_eta.sq TxG_eta.sq
#> 1    R1       0 0.837454 0.864674       0.6438     0.000020   0.000014
#> 2    R2       0 0.074248 0.051435       0.6031     0.001649   0.001964
#> 3    R3       0 0.390655 0.575994       0.6345     0.000353   0.000150
#> 4    R4       0 0.999600 0.916798       0.7003     0.000000   0.000004
#> 5    R5       0 0.602829 0.769201       0.5910     0.000145   0.000046
#> 6    R6       0 0.000774 0.289323       0.5740     0.006267   0.000618
#> 7    R7       0 0.011589 0.047804       0.6005     0.003312   0.002033
#> 8    R8       0 0.757797 0.378064       0.3522     0.000081   0.000661
#> 9    R9       0 0.066243 0.891907       0.4438     0.002459   0.000013
#> 10  R10       0 0.063386 0.201733       0.6645     0.001512   0.000714
#> 11  R11       0 0.268951 0.084972       0.4364     0.000900   0.002188
#> 12  R12       0 0.081867 0.243207       0.5189     0.001906   0.000856
#> 13  R13       0 0.208078 0.420509       0.4722     0.001096   0.000449
#> 14  R14       0 0.703134 0.791410       0.4887     0.000097   0.000047
#> 15  R15       0 0.748125 0.913954       0.5864     0.000056   0.000006
#> 16  R16       0 0.029051 0.865803       0.6582     0.002132   0.000013
#> 17  R17       0 0.854532 0.123018       0.5200     0.000021   0.001497
#> 18  R18       0 0.795859 0.749697       0.4763     0.000046   0.000070
#> 19  R19       0 0.002817 0.007113       0.6413     0.004140   0.003357
#> 20  R20       0 0.012455 0.000484       0.6503     0.002811   0.005503
#> 21  R21       0 0.074750 0.153644       0.3011     0.002901   0.001858
#> 22  R22       0 0.035508 0.496942       0.7029     0.001719   0.000179
#> 23  R23       0 0.293039 0.385011       0.5606     0.000637   0.000435
#> 24  R24       0 0.612788 0.997938       0.6443     0.000120   0.000000
#> 25  R25       0 0.491227 0.476003       0.3498     0.000404   0.000433
#> 26  R26       0 0.752615 0.469005       0.5769     0.000055   0.000291
#> 27  R27       0 0.628675 0.517035       0.7079     0.000090   0.000161
#> 28  R28       0 0.034038 0.445421       0.6262     0.002197   0.000284
#> 29  R29       0 0.044133 0.987922       0.6707     0.001748   0.000000
#> 
#> $Most.uniform.dif
#>   Item Trait_eta.sq Group_eta.sq TxG_eta.sq
#> 6   R6        0.574     0.006267  0.0006184
#> 
#> $Most.non.uniform.dif
#>    Item Trait_eta.sq Group_eta.sq TxG_eta.sq
#> 20  R20       0.6503     0.002811   0.005503
#>