Performs a differential item functioning analysis of polytomous items using ANCOVA

Differential item functioning analysis of polytomous items via ANCOVA

dif.ancova(data, Trait, Group)

Arguments

data

A data frame containing the items. All the items should be scored in the same direction.

Trait

A vector containing measures of the trait. This may be the total scores across all the items, the mean item scores, factor scores, or IRT person measures.

Group

A factor indicating group membership

Examples

library(lordif)
data(Anxiety)
### Perform the dif analysis
out <- dif.ancova(Anxiety[4:32], rowSums(Anxiety[4:32]), Anxiety$gender) 
out
#>     Trait_eta.sq Udif_eta.sq NUdif_eta.sq Trait_p Udif_p NUdif_p
#> R1        0.6438      0.0000       0.0000       0 0.8375  0.8647
#> R2        0.6031      0.0016       0.0020       0 0.0742  0.0514
#> R3        0.6345      0.0004       0.0001       0 0.3907  0.5760
#> R4        0.7003      0.0000       0.0000       0 0.9996  0.9168
#> R5        0.5910      0.0001       0.0000       0 0.6028  0.7692
#> R6        0.5740      0.0063       0.0006       0 0.0008  0.2893
#> R7        0.6005      0.0033       0.0020       0 0.0116  0.0478
#> R8        0.3522      0.0001       0.0007       0 0.7578  0.3781
#> R9        0.4438      0.0025       0.0000       0 0.0662  0.8919
#> R10       0.6645      0.0015       0.0007       0 0.0634  0.2017
#> R11       0.4364      0.0009       0.0022       0 0.2690  0.0850
#> R12       0.5189      0.0019       0.0009       0 0.0819  0.2432
#> R13       0.4722      0.0011       0.0004       0 0.2081  0.4205
#> R14       0.4887      0.0001       0.0000       0 0.7031  0.7914
#> R15       0.5864      0.0001       0.0000       0 0.7481  0.9140
#> R16       0.6582      0.0021       0.0000       0 0.0291  0.8658
#> R17       0.5200      0.0000       0.0015       0 0.8545  0.1230
#> R18       0.4763      0.0000       0.0001       0 0.7959  0.7497
#> R19       0.6413      0.0041       0.0034       0 0.0028  0.0071
#> R20       0.6503      0.0028       0.0055       0 0.0125  0.0005
#> R21       0.3011      0.0029       0.0019       0 0.0747  0.1536
#> R22       0.7029      0.0017       0.0002       0 0.0355  0.4969
#> R23       0.5606      0.0006       0.0004       0 0.2930  0.3850
#> R24       0.6443      0.0001       0.0000       0 0.6128  0.9979
#> R25       0.3498      0.0004       0.0004       0 0.4912  0.4760
#> R26       0.5769      0.0001       0.0003       0 0.7526  0.4690
#> R27       0.7079      0.0001       0.0002       0 0.6287  0.5170
#> R28       0.6262      0.0022       0.0003       0 0.0340  0.4454
#> R29       0.6707      0.0017       0.0000       0 0.0441  0.9879