Gulliksen's item reliability index extended to polytomous items

Takes the item-total correlation and multiplies it with the item standard deviation

Usage

item.reliability.index(x)

Arguments

x

A data frame that contains the items to be analysed

Value

The item reliability index of each item

Examples

item.reliability.index(work_stress[1:9])
#> $rel.index
#>   GWS1   GWS2   GWS3   GWS4   GWS5   GWS6   GWS7   GWS8   GWS9 
#> 0.8454 0.8949 0.8248 0.8421 0.7168 0.6996 0.7969 0.7037 0.7246 
#> 
#> $sorted.rel.index
#>   GWS6   GWS8   GWS5   GWS9   GWS7   GWS3   GWS4   GWS1   GWS2 
#> 0.6996 0.7037 0.7168 0.7246 0.7969 0.8248 0.8421 0.8454 0.8949 
#> 
#> $item.stats
#>         n  raw.r  std.r  r.cor r.drop  mean     sd rel.index
#> GWS1 1377 0.7904 0.7865 0.7535 0.7217 2.601 1.0696    0.8454
#> GWS2 1377 0.7873 0.7794 0.7437 0.7125 2.136 1.1366    0.8949
#> GWS3 1377 0.7914 0.7899 0.7575 0.7250 2.240 1.0423    0.8248
#> GWS4 1377 0.7836 0.7798 0.7440 0.7127 2.224 1.0747    0.8421
#> GWS5 1377 0.7728 0.7779 0.7419 0.7107 2.067 0.9276    0.7168
#> GWS6 1377 0.7706 0.7770 0.7407 0.7095 2.084 0.9079    0.6996
#> GWS7 1377 0.7754 0.7737 0.7368 0.7060 2.680 1.0277    0.7969
#> GWS8 1377 0.7740 0.7799 0.7444 0.7136 1.942 0.9092    0.7037
#> GWS9 1377 0.7591 0.7622 0.7216 0.6918 2.013 0.9546    0.7246
#>