Selecting items based on Gulliksen's item reliability index

Iteratively removes item with the lowest item reliability index from a pool of items. At each step this yields a test with the largest possible variance.

Usage

item.selection(x, nitems)

Arguments

x

A data frame that contains the items to be analysed

nitems

The number of items to retain. This value should never be less than three.

Value

At each step the number of items that have been removed, the name of the last item that has been removed, omega, the average item score, and the standard deviation of the test.

Examples

#Select the combination of k items that will maximise the test variance
item.selection(work_stress[1:9], 5)
#> $Steps
#>   Items.out Items.in Item.removed  Omega  Alpha Average r  Mean     SD
#> 1         0        9         None 0.9185 0.9177    0.5568 2.221 0.7832
#> 2         1        8         GWS6 0.9094 0.9086    0.5571 2.238 0.7969
#> 3         2        7         GWS8 0.8977 0.8969    0.5562 2.280 0.8138
#> 4         3        6         GWS5 0.8831 0.8824    0.5570 2.316 0.8351
#> 5         4        5         GWS9 0.8652 0.8650    0.5626 2.376 0.8629
#> 
#> $Items.in.scale
#> [1] "GWS1" "GWS2" "GWS3" "GWS4" "GWS7"
#> 
#> $Omega
#> $Omega$est
#> [1] 0.8652
#> 
#> 
#> $Alpha
#>  raw_alpha
#>      0.865
#>