Large factor analysis residuals

Identifying large residuals in a matrix of factor analysis residuals

Usage

large.residuals(x, upper = 0.08, lower = -0.08)

Arguments

x

A square matrix of residuals of a factor analysis

upper

Upper limit for acceptable residuals (defaults to 0.08)

lower

Lower limit for acceptable residuals (defaults to -0.)

Value

Statistics relating to the residuals

Examples

library(psychTools)
data(bfi)
myresiduals <- psych::fa(bfi[1:10], nfactors = 2)$resid
myoutput <- large.residuals(x = myresiduals)
#> 
#>   The decimal point is 2 digit(s) to the left of the |
#> 
#>   -8 | 2
#>   -7 | 
#>   -6 | 
#>   -5 | 4
#>   -4 | 4
#>   -3 | 8
#>   -2 | 55
#>   -1 | 980
#>   -0 | 9500
#>    0 | 23456778
#>    1 | 489
#>    2 | 001228
#>    3 | 2245
#>    4 | 3369
#>    5 | 07
#>    6 | 79
#>    7 | 8
#>    8 | 
#>    9 | 59
#> 
myoutput
#> $lower.triangle
#>        A1     A2     A3     A4     A5    C1    C2     C3    C4 C5
#> A1     NA     NA     NA     NA     NA    NA    NA     NA    NA NA
#> A2 -0.082     NA     NA     NA     NA    NA    NA     NA    NA NA
#> A3  0.022 -0.025     NA     NA     NA    NA    NA     NA    NA NA
#> A4  0.028  0.004  0.000     NA     NA    NA    NA     NA    NA NA
#> A5  0.049 -0.025  0.046  0.007     NA    NA    NA     NA    NA NA
#> C1  0.034  0.006  0.019 -0.044  0.032    NA    NA     NA    NA NA
#> C2  0.043  0.003  0.014  0.057 -0.019 0.078    NA     NA    NA NA
#> C3  0.020  0.050 -0.009 -0.038 -0.005 0.005 0.007     NA    NA NA
#> C4  0.095  0.002  0.021  0.035  0.020 0.018 0.032  0.022    NA NA
#> C5  0.000  0.043  0.008 -0.054 -0.010 0.067 0.069 -0.018 0.099 NA
#> 
#> $coordinates
#>    row col
#> A2   2   1
#> C4   9   1
#> C5  10   9
#> 
#> $items
#>   Row Column Cor.residual
#> 2  A2     A1       -0.082
#> 3  C4     A1        0.095
#> 4  C5     C4        0.099
#> 
#> $sorted.residuals
#>  [1] -0.082 -0.054 -0.044 -0.038 -0.025 -0.025 -0.019 -0.018 -0.010 -0.009
#> [11] -0.005  0.000  0.000  0.002  0.003  0.004  0.005  0.006  0.007  0.007
#> [21]  0.008  0.014  0.018  0.019  0.020  0.020  0.021  0.022  0.022  0.028
#> [31]  0.032  0.032  0.034  0.035  0.043  0.043  0.046  0.049  0.050  0.057
#> [41]  0.067  0.069  0.078  0.095  0.099
#> 
#> $median.residual
#> [1] 0.018
#> 
#> $mean.residual
#> [1] 0.016
#> 
#> $max.residual
#> [1] 0.099
#> 
#> $min.residual
#> [1] -0.082
#> 
#> $proportion.large.residuals
#> [1] 0.067
#> 
#> $number.big.residuals
#> [1] 3
#>