Large factor analysis residuals

Identifying large residuals in a matrix of factor analysis residuals

Usage

large.residuals(x, upper = 0.08, lower = -0.08)

Arguments

x

A square matrix of residuals of a factor analysis

upper

Upper limit for acceptable residuals (defaults to 0.08)

lower

Lower limit for acceptable residuals (defaults to -0.)

Value

Statistics relating to the residuals

Examples

# Find the residuals of a factor analysis
myresiduals <- psych::fa(work_stress[1:9], nfactors = 1)$resid

# Identify the pairs of items with the largest residuals
myoutput <- large.residuals(x = myresiduals, upper = -.01, lower = .01)
#> 
#>   The decimal point is 2 digit(s) to the left of the |
#> 
#>   -2 | 32
#>   -1 | 
#>   -1 | 433210
#>   -0 | 999865
#>   -0 | 44321
#>    0 | 00122
#>    0 | 66
#>    1 | 11334
#>    1 | 5579
#>    2 | 
#>    2 | 
#>    3 | 4
#> 
myoutput
#> $lower.triangle
#>        GWS1   GWS2   GWS3   GWS4   GWS5   GWS6   GWS7   GWS8 GWS9
#> GWS1     NA     NA     NA     NA     NA     NA     NA     NA   NA
#> GWS2 -0.001     NA     NA     NA     NA     NA     NA     NA   NA
#> GWS3 -0.014 -0.003     NA     NA     NA     NA     NA     NA   NA
#> GWS4  0.000 -0.006 -0.009     NA     NA     NA     NA     NA   NA
#> GWS5 -0.013  0.002  0.015  0.006     NA     NA     NA     NA   NA
#> GWS6  0.014 -0.023  0.013  0.001 -0.005     NA     NA     NA   NA
#> GWS7  0.034 -0.002 -0.004 -0.004 -0.009  0.000     NA     NA   NA
#> GWS8 -0.009  0.015 -0.010  0.002  0.017 -0.012  0.006     NA   NA
#> GWS9 -0.011  0.019  0.013  0.011 -0.013  0.011 -0.022 -0.008   NA
#> 
#> $coordinates
#>      row col
#> GWS2   2   1
#> GWS3   3   1
#> GWS4   4   1
#> GWS5   5   1
#> GWS6   6   1
#> GWS7   7   1
#> GWS8   8   1
#> GWS9   9   1
#> GWS3   3   2
#> GWS4   4   2
#> GWS5   5   2
#> GWS6   6   2
#> GWS7   7   2
#> GWS8   8   2
#> GWS9   9   2
#> GWS4   4   3
#> GWS5   5   3
#> GWS6   6   3
#> GWS7   7   3
#> GWS8   8   3
#> GWS9   9   3
#> GWS5   5   4
#> GWS6   6   4
#> GWS7   7   4
#> GWS8   8   4
#> GWS9   9   4
#> GWS6   6   5
#> GWS7   7   5
#> GWS8   8   5
#> GWS9   9   5
#> GWS7   7   6
#> GWS8   8   6
#> GWS9   9   6
#> GWS8   8   7
#> GWS9   9   7
#> GWS9   9   8
#> 
#> $items
#>     Row Column Cor.residual
#> 2  GWS2   GWS1       -0.001
#> 3  GWS3   GWS1       -0.014
#> 4  GWS4   GWS1        0.000
#> 5  GWS5   GWS1       -0.013
#> 6  GWS6   GWS1        0.014
#> 7  GWS7   GWS1        0.034
#> 8  GWS8   GWS1       -0.009
#> 9  GWS9   GWS1       -0.011
#> 10 GWS3   GWS2       -0.003
#> 11 GWS4   GWS2       -0.006
#> 12 GWS5   GWS2        0.002
#> 13 GWS6   GWS2       -0.023
#> 14 GWS7   GWS2       -0.002
#> 15 GWS8   GWS2        0.015
#> 16 GWS9   GWS2        0.019
#> 17 GWS4   GWS3       -0.009
#> 18 GWS5   GWS3        0.015
#> 19 GWS6   GWS3        0.013
#> 20 GWS7   GWS3       -0.004
#> 21 GWS8   GWS3       -0.010
#> 22 GWS9   GWS3        0.013
#> 23 GWS5   GWS4        0.006
#> 24 GWS6   GWS4        0.001
#> 25 GWS7   GWS4       -0.004
#> 26 GWS8   GWS4        0.002
#> 27 GWS9   GWS4        0.011
#> 28 GWS6   GWS5       -0.005
#> 29 GWS7   GWS5       -0.009
#> 30 GWS8   GWS5        0.017
#> 31 GWS9   GWS5       -0.013
#> 32 GWS7   GWS6        0.000
#> 33 GWS8   GWS6       -0.012
#> 34 GWS9   GWS6        0.011
#> 35 GWS8   GWS7        0.006
#> 36 GWS9   GWS7       -0.022
#> 37 GWS9   GWS8       -0.008
#> 
#> $sorted.residuals
#>  [1] -0.023 -0.022 -0.014 -0.013 -0.013 -0.012 -0.011 -0.010 -0.009 -0.009
#> [11] -0.009 -0.008 -0.006 -0.005 -0.004 -0.004 -0.003 -0.002 -0.001  0.000
#> [21]  0.000  0.001  0.002  0.002  0.006  0.006  0.011  0.011  0.013  0.013
#> [31]  0.014  0.015  0.015  0.017  0.019  0.034
#> 
#> $median.residual
#> [1] -0.001
#> 
#> $mean.residual
#> [1] 0
#> 
#> $max.residual
#> [1] 0.034
#> 
#> $min.residual
#> [1] -0.023
#> 
#> $proportion.large.residuals
#> [1] 1
#> 
#> $number.big.residuals
#> [1] 36
#>