Reverse code items

Reflects the scoring of items

Usage

reverse.code(x, min.score, max.score, items)

Arguments

x

Data frame containing items, some of which may need to be reverse coded

min.score

Minimum possible score on the items

max.score

Maximum possible score on the items

items

A vector of item names or column numbers of items that need to be reverse coded

Value

A data frame

Examples

library(psychTools)
data(bfi)

## Perform principal components analysis to observe the signs of the component loadings
psych::pca(bfi[1:25], 5)
#> Principal Components Analysis
#> Call: principal(r = r, nfactors = nfactors, residuals = residuals, 
#>     rotate = rotate, n.obs = n.obs, covar = covar, scores = scores, 
#>     missing = missing, impute = impute, oblique.scores = oblique.scores, 
#>     method = method, use = use, cor = cor, correct = 0.5, weight = NULL)
#> Standardized loadings (pattern matrix) based upon correlation matrix
#>      RC2   RC1   RC3   RC5   RC4   h2   u2 com
#> A1  0.16  0.13  0.08 -0.63 -0.13 0.46 0.54 1.4
#> A2  0.03  0.21  0.13  0.71  0.07 0.57 0.43 1.3
#> A3  0.01  0.35  0.09  0.68  0.05 0.59 0.41 1.6
#> A4 -0.06  0.19  0.26  0.53 -0.17 0.41 0.59 2.0
#> A5 -0.12  0.44  0.08  0.56  0.04 0.54 0.46 2.1
#> C1  0.03  0.07  0.65 -0.01  0.21 0.47 0.53 1.2
#> C2  0.11  0.04  0.74  0.10  0.09 0.58 0.42 1.1
#> C3 -0.01  0.00  0.67  0.13 -0.04 0.47 0.53 1.1
#> C4  0.28 -0.04 -0.67 -0.05 -0.12 0.54 0.46 1.4
#> C5  0.33 -0.17 -0.62 -0.05  0.06 0.52 0.48 1.7
#> E1  0.04 -0.68  0.09 -0.07 -0.05 0.48 0.52 1.1
#> E2  0.27 -0.72 -0.07 -0.08 -0.02 0.60 0.40 1.3
#> E3  0.04  0.63  0.07  0.24  0.26 0.53 0.47 1.7
#> E4 -0.10  0.70  0.09  0.28 -0.12 0.61 0.39 1.5
#> E5  0.05  0.58  0.34  0.04  0.20 0.50 0.50 1.9
#> N1  0.79  0.07 -0.04 -0.21 -0.08 0.69 0.31 1.2
#> N2  0.79  0.04 -0.03 -0.19 -0.02 0.66 0.34 1.1
#> N3  0.79 -0.04 -0.06 -0.03  0.00 0.63 0.37 1.0
#> N4  0.65 -0.34 -0.17  0.02  0.09 0.57 0.43 1.7
#> N5  0.63 -0.16 -0.01  0.15 -0.17 0.48 0.52 1.4
#> O1  0.02  0.28  0.13  0.02  0.59 0.44 0.56 1.5
#> O2  0.23  0.03 -0.09  0.11 -0.59 0.43 0.57 1.4
#> O3  0.04  0.36  0.08  0.11  0.64 0.55 0.45 1.7
#> O4  0.28 -0.24 -0.02  0.24  0.50 0.44 0.56 2.6
#> O5  0.12  0.01 -0.04 -0.03 -0.68 0.48 0.52 1.1
#> 
#>                        RC2  RC1  RC3  RC5  RC4
#> SS loadings           3.18 3.09 2.56 2.32 2.11
#> Proportion Var        0.13 0.12 0.10 0.09 0.08
#> Cumulative Var        0.13 0.25 0.35 0.45 0.53
#> Proportion Explained  0.24 0.23 0.19 0.17 0.16
#> Cumulative Proportion 0.24 0.47 0.67 0.84 1.00
#> 
#> Mean item complexity =  1.5
#> Test of the hypothesis that 5 components are sufficient.
#> 
#> The root mean square of the residuals (RMSR) is  0.06 
#>  with the empirical chi square  5469  with prob <  0 
#> 
#> Fit based upon off diagonal values = 0.92
items <- c("A1", "C4", "C5", "E1", "E2", "O2", "O5")
bfi.new <- reverse.code(bfi[1:25], min.score = 1, max.score = 6, items)
#> Error in reverse.code(bfi[1:25], min.score = 1, max.score = 6, items): unused arguments (min.score = 1, max.score = 6)
psych::pca(bfi.new[1:25], 5)
#> Error: object 'bfi.new' not found